Add a menu module. Various compile tweaks to support --board=sitl

+++ src/comms.cpp
@@ -401,7 +401,7 @@ void comms_t::write_status_info_ascii()
 {
     // This info is static so we don't need to send it at a high rate ... once every 10 seconds (?)
     // with an immediate message at the start.
-    console->printf("Uptime: %ld(sec)", AP_HAL::millis() / 1000);
+    console->printf("Uptime: %d(sec)", (unsigned int)(AP_HAL::millis() / 1000));
     console->printf(" SN: %d", config_node.getInt("serial_number"));
     console->printf(" Firmware: %d", FIRMWARE_REV);
     console->printf(" Main loop hz: %d", MASTER_HZ);
